+++ config/routes.rb
@@ -92,7 +92,8 @@ Rails.application.routes.draw do
     end
   end
 
-  get "cookies", to: "cookies#show"
+  get "cookies", to: "cookies#edit"
+  post "cookies", to: "cookies#update"
 
   # High voltage configuration for static pages. Matches routes from the root of the domain. Uses
   # HighVoltage::Constraints::RootRoute to validate that the view exists.