+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add task generators
 - Added a link to the school's information in GIAS on the project information
   page.
+- Users can manage optional cookies.
 
 ## [Release 12][release-12]